requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Deprecated==1.2.14
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
idna==3.8
jcopdl==1.1.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauth==1.0.1
oauthlib==3.2.2
opencv-python==4.10.0.84
openeo==0.31.0
oschmod==0.3.12
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.4
pyperclip==1.9.0
pystac==1.10.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.5.1
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sympy==1.13.2
threadpoolctl==3.5.0
torch==2.4.0+cu124
torchvision==0.19.0+cu124
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
xarray==2024.7.0